requirements.txt
panoramisk==1.4
PyYAML==6.0.2
